Correctly propagate llvm-config errors
Currently, due to the extra `sh`/`cmd` indirection, if the `llvm-config` fails
in the middle of the build script (say in `--libnames`), it will emit an empty
`rustc-link-lib` name because the error is not propagated. The error is also not
shown, and cargo will fail with a mysterious "library name must not be empty".

Remove this extra indirection and improve the error messages by showing the full
context if this happens.

Fixes #45